Update locales.md
I wanted to suggest a change to this page because I couldn't figure out how to add a new value in my controlled list. I finally realized that I had to do two things: update the en.yml file in plugins/local/public/locales/en.yml *and* update the staff frontend for the value to appear.